refactor(sync): extract SyncManager static utilities into dedicated extension files
SyncManager.swift had grown to 2067 lines, 367 over the SwiftLint
file_length warning threshold. Extract self-contained static members
into two focused extension files:

- SyncManager+DateParsing.swift: ISO8601 formatters, regex patterns,
  and date parsing utilities (parseISO8601, truncate/remove helpers,
  addActorMetadata, generateSyncId, extractEntityId)

- SyncManager+ErrorClassification.swift: isAuthenticationError and
  isClerkInfrastructureError static classifiers used by sync retry
  loops to distinguish permanent failures from transient noise

Add // swiftlint:disable file_length to SyncManager.swift following
the same pattern already in use in ProjectorService.swift.

Remove duplicate test-helper stubs in SyncManagerPerformanceTests.swift
that re-implemented parseISO8601/truncate/remove — now accessible
directly since the methods are no longer private.

Lint: 0 violations (196 files). Build and targeted tests pass locally.
