+++ Dequeue/DequeueTests/SyncManagerPerformanceTests.swift
@@ -142,67 +142,5 @@ struct SyncManagerPerformanceTests {
     }
 }
 
-// MARK: - SyncManager Test Helpers
-
-extension SyncManager {
-    /// Exposes parseISO8601 for testing
-    static func parseISO8601(_ string: String) -> Date? {
-        // Use private static method via computed property pattern
-        // This allows testing without making the method public
-        let iso8601WithFractionalSeconds: ISO8601DateFormatter = {
-            let formatter = ISO8601DateFormatter()
-            formatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-            return formatter
-        }()
-
-        let iso8601Standard: ISO8601DateFormatter = {
-            let formatter = ISO8601DateFormatter()
-            formatter.formatOptions = [.withInternetDateTime]
-            return formatter
-        }()
-
-        // First, try parsing as-is with fractional seconds
-        if let date = iso8601WithFractionalSeconds.date(from: string) {
-            return date
-        }
-
-        // If that fails, try truncating nanoseconds to milliseconds
-        let truncated = truncateNanosecondsToMilliseconds(string)
-        if let date = iso8601WithFractionalSeconds.date(from: truncated) {
-            return date
-        }
-
-        // Fall back to standard format without fractional seconds
-        if let date = iso8601Standard.date(from: string) {
-            return date
-        }
-
-        // Last resort: try removing fractional seconds entirely
-        let withoutFractional = removeFractionalSeconds(string)
-        return iso8601Standard.date(from: withoutFractional)
-    }
-
-    /// Exposes truncateNanosecondsToMilliseconds for testing
-    static func truncateNanosecondsToMilliseconds(_ string: String) -> String {
-        // Pattern: (\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})\.(\d{3})\d*(Z|[+-]\d{2}:\d{2})
-        // Captures: 1=datetime, 2=first3digits, 3=timezone
-        let pattern = #"(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})\.(\d{3})\d*(Z|[+-]\d{2}:\d{2})"#
-        guard let regex = try? NSRegularExpression(pattern: pattern) else {
-            return string
-        }
-
-        let range = NSRange(string.startIndex..., in: string)
-        return regex.stringByReplacingMatches(in: string, range: range, withTemplate: "$1.$2$3")
-    }
-
-    /// Exposes removeFractionalSeconds for testing
-    static func removeFractionalSeconds(_ string: String) -> String {
-        let pattern = #"(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2})\.\d+(Z|[+-]\d{2}:\d{2})"#
-        guard let regex = try? NSRegularExpression(pattern: pattern) else {
-            return string
-        }
-
-        let range = NSRange(string.startIndex..., in: string)
-        return regex.stringByReplacingMatches(in: string, range: range, withTemplate: "$1$2")
-    }
-}
+// Note: parseISO8601, truncateNanosecondsToMilliseconds, and removeFractionalSeconds
+// are now accessible directly via SyncManager+DateParsing.swift (no longer private).
